requirements.txt
PyMuPDF
pytesseract
Pillow
pdf2image
nltk
streamlit
google-generativeai
python-dotenv
boto3
awscli